primitives now also inherit from expression
